fix 12

+++ errorhandler.py
@@ -11,7 +11,10 @@ class Errorhandler:
         while True:
             time.sleep(5)
             errorEvent.wait()
-            engine_process.terminate()
+            if engine_process.is_alive():
+                engine_process.terminate()
+            if self.engine_process.is_alive():
+                self.engine_process.terminate()
             time.sleep(1)
             if engine.unstuck_marbles(3):
                 pass
